chore: create a custom hook for gapi
