+++ src/components/Authentication/GoogleAuth/index.jsx
@@ -1,10 +1,8 @@
-import { useEffect } from 'react';
 import { GoogleLogin } from 'react-google-login';
 import PropsTypes from 'prop-types';
-import { gapi } from 'gapi-script';
 import { useNavigate } from 'react-router-dom';
 
-import { useAppDispatch } from 'hooks';
+import { useAppDispatch, useGapi } from 'hooks';
 import { getUser, googleLogin } from 'store/auth/thunk';
 import { message } from 'components/AntDesign';
 import { CLIENT_ID } from 'shared/constants/config';
@@ -12,21 +10,12 @@ import { CLIENT_ID } from 'shared/constants/config';
 export default function GoogleAuth({ label, method }) {
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    const start = () => {
-      gapi.client.init({
-        clientId: CLIENT_ID,
-        scope: 'email',
-      });
-    };
-    gapi.load('client:auth2', start);
-  }, []);
+  useGapi();
 
   const successResponse = async (response) => {
     const { tokenId } = response;
 
-    await dispatch(googleLogin({ method, tokenId })).unwrap();
+    await dispatch(googleLogin({ method, tokenId }));
     navigate('/dashboard');
     dispatch(getUser());
   };
@@ -44,7 +33,6 @@ export default function GoogleAuth({ label, method }) {
       onFailure={failureResponse}
       cookiePolicy="single_host_origin"
     />
-
   );
 }
 GoogleAuth.propTypes = {
