+++ src/hooks/index.js
@@ -1,2 +1,3 @@
 export { default as useAuth } from './useAuth';
 export { useAppDispatch, useAppSelector } from './useRedux';
+export { default as useGapi } from './useGapi';
